✨ Handle `siret_hint`
**Problem**

Some service providers already know the SIRET number they want to
associate with a user.  Currently, users must manually select the
correct SIRET in the ProConnect interface, and it’s difficult for SPs to
indicate which SIRET should be used.

**Proposal**

Expose a `siret_hint` parameter that is used to either prefill the SIRET
field or automatically select the corresponding organization in the
ProConnect interface.