♻ refactor certification dirigeant middleware (#1619)
**Problem**

`checkUserWantToRepresentAnOrganization` redo checks made in previous
user middleware handlers.

`checkUserWantToRepresentAnOrganization` and
`checkUserNeedCertificationDirigeantMiddleware` share common
signification despite checking different information.

`isOrganizationDirigeant` actually perform the Certification Dirigeant
which is not suggested by this name.

It's hard to insert the usage of the siret_hint within the Certification
Dirigeant as this process redirect to organisation selection
independently from the middleware.

**Proposal**

Rename function to explicit what they actually do.

Better insertion of the Ceritification Dirigeant process in the chain
of responsibility in the user middleware.